advent_2025/pyproject.toml
[project]
name = "advent-2025"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "advent-of-code-data>=2.1.0",
    "py-linq==1.4.0",
    "python-stopwatch2==1.1.1",
]
